fixed more compiler warnings
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@13524 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 3fea3270ca03e7ac0965557cb62020ee6f606c6a

+++ motion_planning/sbpl/src/discrete_space_information/nav2d_uu/environment_nav2Duu.cpp
@@ -545,7 +545,7 @@ void EnvironmentNAV2DUU::GetPreds(int stateID, const vector<sbpl_BinaryHiddenVar
 	bool bFreeh = false;
 
 	//iterate over updated h-values to make sure dest cell is not in there
-	for(int hind = 0; hind < updatedhvaluesV->size(); hind++)
+	for(int hind = 0; hind < (int)updatedhvaluesV->size(); hind++)
 	{
 		if(updatedhvaluesV->at(hind).Prob < NAV2DUU_ERR_EPS)
 			bFreeh = true; //there are elements that are free
@@ -602,7 +602,7 @@ void EnvironmentNAV2DUU::GetPreds(int stateID, const vector<sbpl_BinaryHiddenVar
 				//check that it is not known to be free. Otherwise - invalid move
 				int tempID = EnvNAV2DUUCfg.HiddenVariableXY2ID[destx][predY];
 				bool btempfree = false;
-				for(int hind = 0; hind < updatedhvaluesV->size(); hind++)
+				for(int hind = 0; hind < (int)updatedhvaluesV->size(); hind++)
 				{
 					if(updatedhvaluesV->at(hind).h_ID == tempID)
 					{
@@ -625,7 +625,7 @@ void EnvironmentNAV2DUU::GetPreds(int stateID, const vector<sbpl_BinaryHiddenVar
 				//check that it is not known to be free. Otherwise - invalid move
 				int tempID = EnvNAV2DUUCfg.HiddenVariableXY2ID[predX][desty];
 				bool btempfree = false;
-				for(int hind = 0; hind < updatedhvaluesV->size(); hind++)
+				for(int hind = 0; hind < (int)updatedhvaluesV->size(); hind++)
 				{
 					if(updatedhvaluesV->at(hind).h_ID == tempID)
 					{
